Refactor of namespaces and directory structure
